// src/app/pages/label/data-access/api.enum.ts
export const enum EndPointUrl {
    APROPOS_ALL = "apropos/all",
    APROPOS_ADD = "apropos/add",
    APROPOS_UPDATE = "apropos/update",
    APROPOS_DELETE = "apropos/delete",
    APROPOS_ROLES_ALL = "apropos/roles/all",

    CHAMPS_ACTION_ALL = "champs-action/all",
    CHAMPS_ACTION_ADD = "champs-action/add",
    CHAMPS_ACTION_UPDATE = "champs-action/update",
    CHAMPS_ACTION_DELETE = "champs-action/delete",
    CHAMPS_ACTION_ROLES_ALL = "champs-action/roles/all",
}